rename actions object to actionTypes
